debug: Ecs debugger window (#211)
* ui: Add extension shape

* debug: Add skeleton ecs debug panel

* debug: Minor cleanup in physics debugger

* debug: Add component listing

* debug: List archetype count per component

* debug: Show entity count per component

* debug: Add component filtering

* debug: Highlight used components

* debug: Add ecs component sorting

* debug: Fix unstable id's in ecs debugger

* ecs: Store chunk and entity counts as u32

+++ libs/ecs/src/archetype_internal.h
@@ -21,8 +21,8 @@ typedef struct {
   u32    compCount;
   u16*   compOffsetsAndSizes; // u16 offsets[compCount], u16 sizes[compCount].
   void** chunks;              // void* chunks[chunkCount].
-  usize  chunkCount;
-  usize  entityCount;
+  u32    chunkCount;
+  u32    entityCount;
 } EcsArchetype;
 
 ASSERT(sizeof(EcsArchetype) == ecs_archetype_size, "Invalid archetype size");
